requirements.txt
asgiref
Django
django-cors-headers
django-filter
djangorestframework
pillow
sqlparse
tzdata
djangorestframework-simplejwt